feat: responsive

+++ app/Http/Controllers/masterController.php
@@ -10,17 +10,26 @@ use Illuminate\Support\Str;
 
 class masterController extends Controller
 {
-    public function index(){
+    public function index(Request $request){
+        if ($request->user()->cannot('view', User::class)) {
+            abort(403);
+        }
         return view('admin.master.show-user', [
             'user' => User::paginate(10),
         ]);
     }
 
-    public function adduser(){
+    public function adduser(Request $request){
+        if ($request->user()->cannot('view', User::class)) {
+            abort(403);
+        }
         return view('admin.master.add-user');
     }
     public function store(Request $request)
     {
+        if ($request->user()->cannot('create', User::class)) {
+            abort(403);
+        }
         $validateData = $request->validate(
             [
                 'name' => 'required',
@@ -39,13 +48,19 @@ class masterController extends Controller
         }
     }
 
-    public function edit($id)
+    public function edit(Request $request, $id)
     {
+        if ($request->user()->cannot('view', User::class)) {
+            abort(403);
+        }
         $user = User::find($id);
         return view('admin.master.edit-user', ['user' => $user]);
     }
     public function update(Request $request, $id)
     {
+        if ($request->user()->cannot('update', User::class)) {
+            abort(403);
+        }
         $validateData = $request->validate([
             'name' => 'required',
             'email' => 'required|email:dns',
@@ -65,8 +80,11 @@ class masterController extends Controller
         }
     }
 
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
+        if ($request->user()->cannot('delete', User::class)) {
+            abort(403);
+        }
         User::where('id', $id)->delete();
         return redirect('/admin/master')->with('success', 'User Berhasil Dihapus!');
     }
